Compute A Population Projection

Use an initial population pyramid and population change rates to predict future population pyramids.

ComputePopulationProjection(
  initialPopulation,
  populationChangeRates,
  years,
  normalize = NULL,
  growthFlag = TRUE
)

Arguments

initialPopulation

Population structure. The expected format is

List of 4
$ age   : num [1:101] 0 1 2 3 4 5 6 7 8 9 ...
$ female:Formal class 'PopulationPyramid' [package "ehep"] with 2 slots
.. ..@ length: int 101
.. ..@ values: num [1:101] 1739594 1677633 1647809 1617984 1588159 ...
$ male  :Formal class 'PopulationPyramid' [package "ehep"] with 2 slots
.. ..@ length: int 101
.. ..@ values: num [1:101] 1794566 1729423 1697431 1665439 1633447 ...
$ total :Formal class 'PopulationPyramid' [package "ehep"] with 2 slots
.. ..@ length: int 101
.. ..@ values: num [1:101] 3534160 3407056 3345240 3283423 3221606 ...

populationChangeRates

Population change rates (both fertility and mortality)

years

Vector of years to model

normalize

Whether or not to normalize the initial population default = NULL, meaning don't normalize. A numeric value means normalize to that value.

growthFlag

If FALSE, normalize each year to the same population as the initial year (default = TRUE)

Value

Demographics time-series

Examples

if (FALSE) {
library(ehep)
ehep::Trace(TRUE)

ehep::InitializePopulation()
ehep::InitializeHealthcareTasks()
ehep::InitializeScenarios()
ehep::InitializeStochasticParameters()
ehep::InitializeSeasonality()

scenario <- "ScenarioName"

set.seed(54321)

scenarioData <- SaveBaseSettings(scenario)
ConfigureExperimentValues()

exp <- ehep:::EXP
gpe <- ehep:::GPE

population <- ComputePopulationProjection(
  exp$initialPopulation,
  exp$populationChangeRates,
  gpe$years,
  normalize = scenarioData$BaselinePop,
  growthFlag = scenarioData$o_PopGrowth
)
}